fixed last 2 warnings on linux in SILACAnalyzer.C; some cosmetics in MapAlignmentAlgorithmSpectrumAlignment OpenMS should be FREE of warnings now. keep it that way.
git-svn-id: https://open-ms.svn.sourceforge.net/svnroot/open-ms/branches/Release1.4@5028 6adb6e08-d915-0410-941f-83917bcadc18

+++ source/APPLICATIONS/TOPP/SILACAnalyzer.C
@@ -545,7 +545,7 @@ class TOPPSILACAnalyzer
 						UInt j=0;
 						for (std::vector<SILACData>::iterator it=data.begin(); it!= data.end(); ++it)
 						{
-							if ((UInt)it->cluster_id==i)
+							if (it->cluster_id==i)
 							{
 								i1[3*j] = it->int1;
 								i2[3*j] = it->int4;
@@ -697,7 +697,7 @@ class TOPPSILACAnalyzer
 						UInt j=0;
 		   			for (std::vector<SILACData>::iterator it=data.begin(); it!= data.end(); ++it)
 						{
-							if ((UInt)it->cluster_id==i)
+							if (it->cluster_id==i)
 							{
 								i1[3*j] = it->int1;
 								i2[3*j] = it->int4;
